rename page settings
